// src/backend/marketing-cloud/models/rfq-journey-entry.dto.ts
export class RFQJourneyEntryDTO {
  Email: string;
  Phone: string;
  LastName: string;
  FirstName: string;
  CompanyName: string;
  DeliveryZIP: string;
  DeliveryState: string;
  StartDate: Date;
  EndDate: Date;
  KindofUse: string;
  RentalPurpose: string;
  RentalDuration: string;
  StandardProducts: string;
  SpecialtyProducts: string;
  Acq_Company: string;
  crm_web_lead_privacy_policy_opt_in: boolean;
  CustomerType: string;
  WebRFQId: string;
  LastClickId: string;
}